Fleshed out menu code, edited asset
Menu now accepts arbitrary input in the form of lists of strings. It has
the ability to facilitate executing arbitrary code for each selection by
returning an index to the calling object.
Modified BMPText so it can change fonts on the fly (done mainly to test
menu code).
Edited the font bit map as some of the letters were offset for some
reason.