Fleshed out inputcfg some more
Made inputcfg functional, but had to scale back the integration since
managing input events from within the class was proving to be too
puzzling. The current solution therefore was to utilize conditionals
outside of the class (in the test.cpp loop) to determine when to wait
for the user response to a key assignment.