Implemented Spelunky style walk direction change
If move_left is held down, then move_right is pressed, the player should
stop facing left. As soon as one of the keys is released, the player
should start moving in the direction of the key that is still held down.
To affect this I had to add another getter method to inputmap.