cleaned up sprite & text drawing routines
Eliminated the need to pass in the sprite/text dimension as parameter to
sprite draw and by extension, text print, functions. Also needed to
modify txtlayer receive text function to accommodate this. A lot of
unnecessary code was removed;after all these are bitmapped fonts we're
talking about so the size of the sprite will determine the size of the
font.