moved anim data to files also
Animation data also lives in a text file, with format similar to the
spritesheet file. There is less info in the file since more of the info
is able to be calculated. Less stuff to type in the better!
Also disabled backNforth for the run left animation to check and see if
that parameter gets passed.