reworked txtlayer system
Tried to get clever and treat txtlayer as a dumb dump for any and all
text happening anytime, but because of the nature of how it works (like
the renderer, wipes and redraws every frame) this caused strings in
loops that are called intermittently to show up...intermittently.
Ultimately I had to go back to the old way of filling up a separate
string variable with the desired text, and waiting until the base level
of the main loop before tossing the messages in the txtlayer.
I'm not too concerned about making this elegant at the moment. time to
move on...