Changed timer
Now there's an internal timekeeper variable. It is an integer for
stability over time. Thus it keeps the value straight from
SDL_GetPerformanceCounter instead of the actual time in seconds. The
time is derived when needed as usual with SDL_GetPerformanceFrequency.

Currently there seems to be an issue with getting the proper frameTime
so I've added a hard value for now.