Added two axes of overlap detection
Added overlap detection for both X and Y axes. For some reason
everything was peachy when just the X was there, but now things are
wonky with the addition of the Y stuff, which in theory should be
completely independent from the X stuff at least where endpoints are
concerned.